Move functions and models into optimization app
