+++ offgridplanner/optimization/grid/identify_consumers_on_map.py
