--- offgridplanner/steps/helpers.py
@@ -1,2 +1,0 @@
-def is_ajax(request):
-    return request.headers.get("x-requested-with") == "XMLHttpRequest"
