+++ offgridplanner/templates/pages/demand_estimation.html
@@ -12,7 +12,7 @@
 
 {% block start_body_scripts %}
 <script>
-  const loadDemandPlotUrl = `{% url 'projects:load_demand_plot_data' proj_id %}`;
+  const loadDemandPlotUrl = `{% url 'optimization:load_demand_plot_data' proj_id %}`;
 </script>
 {% endblock start_body_scripts %}
