+++ offgridplanner/templates/pages/simulation_results.html
@@ -568,9 +568,10 @@
 {% block end_body_scripts %}
     <script>
       const proj_id = '{{ proj_id }}';
-      const loadResultsUrl = `{% url 'projects:load_results' proj_id %}`;
-      const dbNodesToJsUrl = `{% url 'projects:db_nodes_to_js' proj_id %}`;
-      const dbLinksToJsUrl = `{% url 'projects:db_links_to_js' proj_id %}`;
+      const loadResultsUrl = `{% url 'optimization:load_results' proj_id %}`;
+      const dbNodesToJsUrl = `{% url 'optimization:db_nodes_to_js' proj_id %}`;
+      const dbLinksToJsUrl = `{% url 'optimization:db_links_to_js' proj_id %}`;
+      const loadPlotDataUrl = `{% url 'optimization:load_plot_data' proj_id %}`;
     </script>
   <script src="{% static 'js/integrate_map.js' %}"></script>
   <script src="https://d3js.org/d3.v3.min.js" charset="utf-8"></script>
